resolved impuation from silver to gold, issue 13
