+++ src/ml_engineering/model_preprocess.py
@@ -106,95 +106,4 @@ def validate_master_dataset(
     return df
 
 
-def _identify_ohe_columns(df: pd.DataFrame) -> List[str]:
-    """Return columns that are binary OHE flags (values exclusively in {0, 1, NaN})."""
-    ohe_cols = []
-    for col in df.select_dtypes(include=[np.number]).columns:
-        unique_vals = df[col].dropna().unique()
-        if len(unique_vals) <= 2 and set(unique_vals).issubset({0.0, 1.0}):
-            ohe_cols.append(col)
-    return ohe_cols
 
-
-def impute_missing_values(
-    df: pd.DataFrame,
-    *,
-    numeric_strategy: str = "median",
-    add_missing_indicator: bool = True,
-) -> pd.DataFrame:
-    """
-    Return a fully imputed copy of *df* with all columns cast to float64.
-
-    Imputation tiers:
-    - Binary OHE flags ({0, 1, NaN}) → filled with 0 (absent = not observed)
-    - Remaining numeric columns → SimpleImputer(strategy=numeric_strategy)
-    - Non-numeric columns → dropped (should not exist in a Gold table)
-
-    Args:
-        df: Input DataFrame (typically master_data_ml_joined).
-        numeric_strategy: 'median' (default, robust to skew) or 'mean'.
-        add_missing_indicator: If True, creates <col>_is_missing binary flags.
-
-    Returns:
-        A new DataFrame with zero NaNs, all float64.
-    """
-    result = df.copy()
-    f_log(f"Starting imputation. Input shape: {result.shape}, "
-          f"total NaNs: {result.isna().sum().sum()}", c_type="process")
-
-    # 0. Extract structural keys — these are the dataset's spine, not features
-    structural_keys = [c for c in [DATE_COL, SBI_COL] if c in result.columns]
-    key_data = result[structural_keys].copy() if structural_keys else None
-    result = result.drop(columns=structural_keys, errors="ignore")
-    f_log(f"Preserved {len(structural_keys)} structural keys: {structural_keys}", c_type="info")
-
-    # 1. Capture missing-indicator flags BEFORE any imputation
-    if add_missing_indicator:
-        cols_with_nans = result.columns[result.isna().any()].tolist()
-        missing_flags = result[cols_with_nans].isna().astype(float)
-        missing_flags.columns = [f"{c}_is_missing" for c in cols_with_nans]
-        f_log(f"Created {len(cols_with_nans)} missing-indicator columns.", c_type="info")
-
-    # 2. Drop non-numeric columns (should already be absent in Gold, but safety net)
-    non_numeric_cols = result.select_dtypes(exclude=[np.number]).columns.tolist()
-    if non_numeric_cols:
-        f_log(f"Dropping {len(non_numeric_cols)} non-numeric columns: "
-              f"{non_numeric_cols[:5]}{'...' if len(non_numeric_cols) > 5 else ''}", c_type="warning")
-        result = result.drop(columns=non_numeric_cols)
-
-    # 3. Separate OHE flags from continuous numeric columns
-    ohe_cols = _identify_ohe_columns(result)
-    numeric_cols = [c for c in result.columns if c not in ohe_cols]
-
-    # 4. Impute OHE flags with 0 (absent = not observed)
-    if ohe_cols:
-        result[ohe_cols] = result[ohe_cols].fillna(0.0)
-        f_log(f"Imputed {len(ohe_cols)} OHE columns with 0.", c_type="info")
-
-    # 5. Impute continuous numeric columns with median/mean
-    cols_needing_imputation = [c for c in numeric_cols if result[c].isna().any()]
-    if cols_needing_imputation:
-        imputer = SimpleImputer(strategy=numeric_strategy)
-        result[cols_needing_imputation] = imputer.fit_transform(
-            result[cols_needing_imputation]
-        )
-        f_log(f"Imputed {len(cols_needing_imputation)} numeric columns "
-              f"with {numeric_strategy}.", c_type="info")
-
-    # 6. Append missing-indicator columns
-    if add_missing_indicator and len(cols_with_nans) > 0:
-        # Only keep indicators for columns that survived the drop step
-        surviving_indicators = [c for c in missing_flags.columns
-                                if c.replace("_is_missing", "") in result.columns]
-        result = pd.concat([result, missing_flags[surviving_indicators]], axis=1)
-
-    # 7. Cast feature columns to float64 (MLflow / Gold Policy)
-    result = result.astype("float64")
-
-    # 8. Re-attach structural keys at the front of the DataFrame
-    if key_data is not None:
-        result = pd.concat([key_data.reset_index(drop=True), result.reset_index(drop=True)], axis=1)
-
-    f_log(f"Imputation complete. Output shape: {result.shape}, "
-          f"remaining NaNs: {result.isna().sum().sum()}", c_type="success")
-    return result
