+++ tests/test_model_preprocess.py
@@ -9,8 +9,8 @@ import unittest
 import numpy as np
 import pandas as pd
 
-from src.ml_engineering.model_preprocess import (
-    validate_master_dataset,
+from src.ml_engineering.model_preprocess import validate_master_dataset
+from src.utils.m_imputation import (
     impute_missing_values,
     _identify_ohe_columns,
 )
